covid_sync/requirements.txt
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
django